Move DataSourceEnumContract to Contracts and rename
Renamed DataSourceEnumContract to DataSourceEnum and moved it from DataSources to Contracts namespace. Updated all references to use the new interface and namespace, improving code organization and consistency.
